# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Brilliant Sidecar API
  version: 1.4.0
  description: |
    REST API for interacting with cached LMS data and Model Context Protocol (MCP) server integration.
    
    All `/api/v1/*` endpoints require authentication via:
    1. `X-API-KEY` header
    2. Bearer JWT Token in `Authorization` header
    3. Active internal session (for browser-based SPA)
servers:
  - url: /api/v1
    description: Local API server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Notification:
      type: object
      properties:
        id: { type: integer }
        external_id: { type: string }
        title: { type: string }
        body: { type: string }
        date: { type: string, format: date-time }
        course_id: { type: string }
        course_name: { type: string }
        semester: { type: string }
        urgency: { type: integer, description: "1 (Low) to 5 (High)" }
        is_personal: { type: boolean }
        is_read: { type: boolean }
        url: { type: string }
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /calendar:
    get:
      summary: Get items for calendar visualization
      parameters:
        - name: view
          in: query
          schema: { type: string, enum: [month, week] }
        - name: date
          in: query
          schema: { type: string, format: date }
        - name: show_completed
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Grouped items with visualization metadata
          content:
            application/json:
              schema:
                type: object
...
  /notifications:
    get:
      summary: Retrieve notifications with filtering
      parameters:
        - name: course_id
          in: query
          schema: { type: string }
        - name: semester
          in: query
          schema: { type: string }
        - name: urgency
          in: query
          schema: { type: integer }
        - name: is_personal
          in: query
          schema: { type: boolean }
        - name: show_read
          in: query
          description: Include read notifications (default false)
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
...